+++ ios/Classes/AutomationsPlugin.swift
@@ -73,8 +73,7 @@ public class AutomationsPlugin: NSObject {
 
   public func showScreen(_ screenId: String?, _ result: @escaping FlutterResult) {
     guard let screenId = screenId else {
-      result(FlutterError.noNecessaryData)
-      return
+      return result(FlutterError.noNecessaryData)
     }
 
     automationSandwich.showScreen(screenId) { data, error in
@@ -84,6 +83,16 @@ public class AutomationsPlugin: NSObject {
       
       result(data)
     }
+  }
+
+  public func setScreenPresentationConfig(_ configData: [String: Any]?, _ screenId: String?,  _ result: @escaping FlutterResult) {
+    guard let configData = configData else {
+      result(FlutterError.noNecessaryData)
+      return
+    }
+
+    automationSandwich.setScreenPresentationConfig(configData, forScreenId:screenId)
+    result(nil)
   }
 }