+++ ios/Classes/SwiftQonversionPlugin.swift
@@ -135,6 +135,10 @@ public class SwiftQonversionPlugin: NSObject, FlutterPlugin {
       automationsPlugin?.showScreen(args["screenId"] as? String, result)
       return
 
+    case "setScreenPresentationConfig":
+      automationsPlugin?.setScreenPresentationConfig(args["configData"] as? [String: Any], args["screenId"] as? String, result)
+      return
+
     default:
       return result(FlutterMethodNotImplemented)
     }
@@ -149,12 +153,14 @@ public class SwiftQonversionPlugin: NSObject, FlutterPlugin {
     }
     let environment = args["environment"] as? String
     let entitlementsCacheLifetime = args["entitlementsCacheLifetime"] as? String
+    let proxyUrl = args["proxyUrl"] as? String
 
     qonversionSandwich?.initialize(
       projectKey: projectKey,
       launchModeKey: launchMode,
       environmentKey: environment,
-      entitlementsCacheLifetimeKey: entitlementsCacheLifetime
+      entitlementsCacheLifetimeKey: entitlementsCacheLifetime,
+      proxyUrl: proxyUrl
     )
   }