+++ lib/src/internal/constants.dart
@@ -4,6 +4,7 @@ class Constants {
   static const kLaunchMode = 'launchMode';
   static const kEnvironment = 'environment';
   static const kEntitlementsCacheLifetime = 'entitlementsCacheLifetime';
+  static const kProxyUrl = 'proxyUrl';
   static const kUserId = 'userId';
   static const kData = 'data';
   static const kProvider = 'provider';
@@ -22,6 +23,7 @@ class Constants {
   static const kNotificationData = 'notificationData';
   static const kLifetime = 'lifetime';
   static const kScreenId = 'screenId';
+  static const kConfigData = 'configData';
 
   // MethodChannel methods names
   static const mInitialize = 'initialize';
@@ -53,6 +55,7 @@ class Constants {
   static const mHandleNotification = 'automationsHandleNotification';
   static const mGetNotificationCustomPayload = 'automationsGetNotificationCustomPayload';
   static const mShowScreen = 'automationsShowScreen';
+  static const mSetScreenPresentationConfig = 'setScreenPresentationConfig';
 
   // Numeric constants
   static const skuDetailsPriceRatio = 1000000;