+++ macos/Classes/SwiftQonversionPlugin.swift
@@ -111,12 +111,14 @@ public class SwiftQonversionPlugin: NSObject, FlutterPlugin {
     }
     let environment = args["environment"] as? String
     let entitlementsCacheLifetime = args["entitlementsCacheLifetime"] as? String
+    let proxyUrl = args["proxyUrl"] as? String
 
     qonversionSandwich?.initialize(
       projectKey: projectKey,
       launchModeKey: launchMode,
       environmentKey: environment,
-      entitlementsCacheLifetimeKey: entitlementsCacheLifetime
+      entitlementsCacheLifetimeKey: entitlementsCacheLifetime,
+      proxyUrl: proxyUrl
     )
   }